wip(article): Fix styles, looks good, remove unused libs, remove Article tools
