+++ package.json
@@ -76,6 +76,7 @@
     "@types/crypto-js": "^4.2.2",
     "@types/leaflet": "^1.9.12",
     "@types/leaflet.markercluster": "^1.5.4",
+    "@types/markdown-it-emoji": "^3.0.1",
     "@types/papaparse": "^5.3.14",
     "@types/pluralize": "^0.0.33",
     "assert": "^2.1.0",
@@ -121,6 +122,7 @@
     "long": "^5.2.1",
     "lottie-react-native": "6.5.1",
     "markdown-it": "^14.1.0",
+    "markdown-it-emoji": "^3.0.0",
     "merkletreejs": "^0.4.0",
     "metamask-react": "^2.4.1",
     "moment": "^2.29.4",
@@ -129,7 +131,6 @@
     "papaparse": "^5.4.1",
     "plausible-tracker": "^0.3.8",
     "pluralize": "^8.0.0",
-    "prop-types": "^15.8.1",
     "protobufjs": "^7.2.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -142,7 +143,6 @@
     "react-native-confetti-cannon": "^1.5.2",
     "react-native-countdown-component": "^2.7.1",
     "react-native-drax": "^0.10.2",
-    "react-native-fit-image": "^1.5.5",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "~2.14.0",
     "react-native-get-random-values": "~1.8.0",
