+++ packages/components/navigation/getNormalModeScreens.tsx
@@ -9,7 +9,6 @@ import { CoreDAOScreen } from "@/screens/CoreDAO/CoreDAOScreen";
 import { DAppStoreScreen } from "@/screens/DAppStore/DAppStoreScreen";
 import { ToriPunks } from "@/screens/DAppStore/apps/toripunks/HomeScreen";
 import { FeedScreen } from "@/screens/Feed/FeedScreen";
-import { FeedNewArticleMarkdownScreen } from "@/screens/FeedNewArticle/FeedNewArticleMarkdownScreen.tsx";
 import { FeedNewArticleScreen } from "@/screens/FeedNewArticle/FeedNewArticleScreen.tsx";
 import { FeedPostViewScreen } from "@/screens/FeedPostView/FeedPostViewScreen";
 import { GovernanceProposalScreen } from "@/screens/Governance/GovernanceProposal/GovernanceProposalScreen";
@@ -452,11 +451,6 @@ export const getNormalModeScreens = ({
       <Nav.Screen
         name="FeedNewArticle"
         component={FeedNewArticleScreen}
-        options={{ header: () => null, title: screenTitle("New Article") }}
-      />
-      <Nav.Screen
-        name="FeedNewArticleMarkdown"
-        component={FeedNewArticleMarkdownScreen}
         options={{ header: () => null, title: screenTitle("New Article") }}
       />
