+++ packages/screens/FeedNewArticle/FeedNewArticleScreen.tsx
@@ -1,23 +1,24 @@
 import pluralize from "pluralize";
-import React, { useEffect, useRef, useState } from "react";
-import { Controller, useForm } from "react-hook-form";
-import { ScrollView, View } from "react-native";
+import { useEffect, useRef, useState } from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import { ScrollView, useWindowDimensions, View } from "react-native";
 import { useSelector } from "react-redux";
 
-import priceSVG from "../../../assets/icons/price.svg";
 import useSelectedWallet from "../../hooks/useSelectedWallet";
 
+import penSVG from "@/assets/icons/pen.svg";
+import priceSVG from "@/assets/icons/price.svg";
 import { BrandText } from "@/components/BrandText";
 import { SVG } from "@/components/SVG";
 import { ScreenContainer } from "@/components/ScreenContainer";
-import { ScreenTitle } from "@/components/ScreenContainer/ScreenTitle";
 import { WalletStatusBox } from "@/components/WalletStatusBox";
 import { TertiaryBox } from "@/components/boxes/TertiaryBox";
+import { CustomPressable } from "@/components/buttons/CustomPressable";
 import { Label, TextInputCustom } from "@/components/inputs/TextInputCustom";
 import { FileUploader } from "@/components/inputs/fileUploader";
 import { FeedPostingProgressBar } from "@/components/loaders/FeedPostingProgressBar";
-import { RichText } from "@/components/socialFeed/RichText";
 import { PublishValues } from "@/components/socialFeed/RichText/RichText.type";
+import { SocialArticleCard } from "@/components/socialFeed/SocialCard/cards/SocialArticleCard";
 import { MapModal } from "@/components/socialFeed/modals/MapModal/MapModal";
 import { SpacerColumn } from "@/components/spacer";
 import { useFeedbacks } from "@/context/FeedbacksProvider";
@@ -25,8 +26,10 @@ import { useWalletControl } from "@/context/WalletControlProvider";
 import { useFeedPosting } from "@/hooks/feed/useFeedPosting";
 import { useIpfs } from "@/hooks/useIpfs";
 import { useIsMobile } from "@/hooks/useIsMobile";
+import { useMaxResolution } from "@/hooks/useMaxResolution";
 import { useSelectedNetworkId } from "@/hooks/useSelectedNetwork";
 import { NetworkFeature } from "@/networks";
+import { ArticleContentEditor } from "@/screens/FeedNewArticle/components/ArticleContentEditor/ArticleContentEditor";
 import { selectNFTStorageAPI } from "@/store/slices/settings";
 import { feedPostingStep, FeedPostingStepId } from "@/utils/feed/posting";
 import { generateArticleMetadata } from "@/utils/feed/queries";
@@ -34,29 +37,33 @@ import { generateIpfsKey } from "@/utils/ipfs";
 import { IMAGE_MIME_TYPES } from "@/utils/mime";
 import { ScreenFC, useAppNavigation } from "@/utils/navigation";
 import {
-  ARTICLE_COVER_IMAGE_MAX_HEIGHT,
-  ARTICLE_COVER_IMAGE_RATIO,
-  ARTICLE_THUMBNAIL_IMAGE_MAX_HEIGHT,
-  ARTICLE_THUMBNAIL_IMAGE_MAX_WIDTH,
-} from "@/utils/social-feed";
-import {
   neutral00,
   neutral11,
+  neutral33,
   neutral77,
+  neutralFF,
   secondaryColor,
 } from "@/utils/style/colors";
-import { fontSemibold13 } from "@/utils/style/fonts";
-import { layout, screenContentMaxWidth } from "@/utils/style/layout";
+import { fontSemibold13, fontSemibold20 } from "@/utils/style/fonts";
+import {
+  layout,
+  RESPONSIVE_BREAKPOINT_S,
+  screenContentMaxWidth,
+} from "@/utils/style/layout";
 import {
   CustomLatLngExpression,
   NewArticleFormValues,
   PostCategory,
+  SocialFeedArticleMetadata,
 } from "@/utils/types/feed";
 import { RemoteFileData } from "@/utils/types/files";
 
 //TODO: In mobile : Make ActionsContainer accessible (floating button ?)
 
 export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
+  const { width } = useMaxResolution();
+  const { width: windowWidth } = useWindowDimensions();
+  const isSmallScreen = windowWidth < RESPONSIVE_BREAKPOINT_S;
   const isMobile = useIsMobile();
   const wallet = useSelectedWallet();
   const selectedNetworkId = useSelectedNetworkId();
@@ -87,7 +94,7 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
         message: "",
       });
       navigateBack();
-      reset();
+      newArticleForm.reset();
     }, 1000);
   });
   const forceNetworkFeature = NetworkFeature.SocialFeed;
@@ -97,15 +104,15 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
   const { setToast } = useFeedbacks();
   const navigation = useAppNavigation();
   const scrollViewRef = useRef<ScrollView>(null);
+  const [isThumbnailButtonHovered, setThumbnailButtonHovered] = useState(false);
   const [location, setLocation] = useState<CustomLatLngExpression>();
   const [isMapShown, setIsMapShown] = useState(false);
-  const {
-    control,
-    setValue,
-    reset,
-    watch,
-    formState: { errors },
-  } = useForm<NewArticleFormValues>({
+  const cardStyle = isSmallScreen && {
+    borderRadius: 0,
+    borderLeftWidth: 0,
+    borderRightWidth: 0,
+  };
+  const newArticleForm = useForm<NewArticleFormValues>({
     defaultValues: {
       title: "",
       message: "",
@@ -118,13 +125,15 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
     },
     mode: "onBlur",
   });
-  //TODO: Not handled for now
-  // const { mutate: openGraphMutate, data: openGraphData } = useOpenGraph();
-
-  const formValues = watch();
-
-  //TODO: Keep short post formValues when returning to short post
-  const navigateBack = () => navigation.navigate("Feed");
+  const formValues = newArticleForm.watch();
+  const previewMetadata: SocialFeedArticleMetadata = {
+    title: formValues.title,
+    shortDescription: formValues.shortDescription || "",
+    thumbnailImage: formValues.thumbnailImage,
+    message: "",
+    hashtags: [],
+    mentions: [],
+  };
 
   const onPublish = async (values: PublishValues) => {
     const action = "Publish an Article";
@@ -227,11 +236,11 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
     }
   };
 
-  // Scroll to bottom when the loading bar appears
-  useEffect(() => {
-    if (step.id !== "UNDEFINED" && isLoading)
-      scrollViewRef.current?.scrollToEnd();
-  }, [step, isLoading]);
+  //TODO: Keep short post formValues when returning to short post
+  const navigateBack = () => navigation.navigate("Feed");
+
+  //TODO: Not handled for now
+  // const { mutate: openGraphMutate, data: openGraphData } = useOpenGraph();
 
   // // OpenGraph URL preview
   // useEffect(() => {
@@ -243,21 +252,29 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
   //   })
   // }, [addedUrls])
 
+  // Scroll to bottom when the loading bar appears
+  useEffect(() => {
+    if (step.id !== "UNDEFINED" && isLoading)
+      scrollViewRef.current?.scrollToEnd();
+  }, [step, isLoading]);
+
   return (
     <ScreenContainer
       forceNetworkFeatures={[forceNetworkFeature]}
       responsive
       mobileTitle="NEW ARTICLE"
       fullWidth
-      headerChildren={<ScreenTitle>New Article</ScreenTitle>}
+      headerChildren={<BrandText style={fontSemibold20}>New Article</BrandText>}
       onBackPress={navigateBack}
       footerChildren
+      noMargin
       noScroll
     >
       <ScrollView
         ref={scrollViewRef}
         style={{
-          marginTop: isMobile ? layout.spacing_x2 : layout.contentSpacing,
+          paddingTop: isMobile ? layout.spacing_x2 : layout.contentSpacing,
+          paddingBottom: layout.spacing_x2,
         }}
         contentContainerStyle={{
           width: "100%",
@@ -265,145 +282,160 @@ export const FeedNewArticleScreen: ScreenFC<"FeedNewArticle"> = () => {
           alignSelf: "center",
         }}
       >
-        <WalletStatusBox />
-        <SpacerColumn size={3} />
-
-        <TertiaryBox
+        <View
           style={{
-            paddingVertical: layout.spacing_x1,
-            paddingHorizontal: layout.spacing_x1_5,
-            flexDirection: "row",
-            alignItems: "center",
-            justifyContent: "flex-start",
-            height: 48,
-            backgroundColor: neutral11,
+            alignSelf: "center",
+            width,
           }}
         >
-          <SVG
-            source={priceSVG}
-            height={24}
-            width={24}
-            color={secondaryColor}
-          />
-          <BrandText
-            style={[
-              fontSemibold13,
-              { color: neutral77, marginLeft: layout.spacing_x1 },
-            ]}
-          >
-            {freePostCount
-              ? `You have ${freePostCount} free ${pluralize(
-                  "Article",
-                  freePostCount,
-                )} left`
-              : `The cost for this Article is ${prettyPublishingFee}`}
-          </BrandText>
-        </TertiaryBox>
+          <WalletStatusBox />
+          <SpacerColumn size={3} />
 
-        <FileUploader
-          label="Thumbnail image"
-          style={{
-            marginTop: layout.spacing_x3,
-          }}
-          fileImageStyle={{
-            objectFit: "cover",
-            height: ARTICLE_THUMBNAIL_IMAGE_MAX_HEIGHT,
-            maxWidth: ARTICLE_THUMBNAIL_IMAGE_MAX_WIDTH,
-          }}
-          onUpload={(files) =>
-            setValue("thumbnailImage", {
-              isThumbnailImage: true,
-              ...files[0],
-            })
-          }
-          mimeTypes={IMAGE_MIME_TYPES}
-        />
-
-        <FileUploader
-          label="Cover image"
-          style={{
-            marginTop: layout.spacing_x3,
-            width: "100%",
-          }}
-          fileImageStyle={{
-            objectFit: "cover",
-            height: "100%",
-            width: "100%",
-            maxHeight: ARTICLE_COVER_IMAGE_MAX_HEIGHT,
-            aspectRatio: ARTICLE_COVER_IMAGE_RATIO,
-          }}
-          onUpload={(files) =>
-            setValue("coverImage", {
-              isCoverImage: true,
-              ...files[0],
-            })
-          }
-          mimeTypes={IMAGE_MIME_TYPES}
-        />
-
-        <TextInputCustom<NewArticleFormValues>
-          noBrokenCorners
-          rules={{ required: true }}
-          height={48}
-          label="Title"
-          placeHolder="Type title here"
-          name="title"
-          control={control}
-          variant="labelOutside"
-          containerStyle={{ marginVertical: layout.spacing_x3 }}
-          boxMainContainerStyle={{
-            backgroundColor: neutral00,
-            borderRadius: 12,
-          }}
-        />
-
-        <TextInputCustom<NewArticleFormValues>
-          noBrokenCorners
-          rules={{ required: true }}
-          multiline
-          label="Short description"
-          placeHolder="Type short description here"
-          name="shortDescription"
-          control={control}
-          variant="labelOutside"
-          containerStyle={{ marginBottom: layout.spacing_x3 }}
-          boxMainContainerStyle={{
-            backgroundColor: neutral00,
-            borderRadius: 12,
-          }}
-        />
-
-        <View>
-          <Label>Article content</Label>
-          <SpacerColumn size={1} />
-          <Controller
-            name="message"
-            control={control}
-            rules={{
-              required: true,
+          <TertiaryBox
+            style={{
+              paddingVertical: layout.spacing_x1,
+              paddingHorizontal: layout.spacing_x1_5,
+              flexDirection: "row",
+              alignItems: "center",
+              justifyContent: "flex-start",
+              height: 48,
+              backgroundColor: neutral11,
             }}
-            render={({ field: { onChange, onBlur } }) => (
-              <RichText
-                onChange={onChange}
-                onBlur={onBlur}
-                initialValue={formValues.message}
-                loading={isLoading}
-                publishDisabled={
-                  errors?.message?.type === "required" ||
-                  !formValues.message ||
-                  !formValues.title ||
-                  !formValues.shortDescription ||
-                  !wallet
-                }
-                onPublish={onPublish}
-                authorId={userId || ""}
-                postId=""
-                setIsMapShown={setIsMapShown}
-                hasLocation={!!location}
-              />
-            )}
+          >
+            <SVG
+              source={priceSVG}
+              height={24}
+              width={24}
+              color={secondaryColor}
+            />
+            <BrandText
+              style={[
+                fontSemibold13,
+                { color: neutral77, marginLeft: layout.spacing_x1 },
+              ]}
+            >
+              {freePostCount
+                ? `You have ${freePostCount} free ${pluralize(
+                    "Article",
+                    freePostCount,
+                  )} left`
+                : `The cost for this Article is ${prettyPublishingFee}`}
+            </BrandText>
+          </TertiaryBox>
+
+          <TextInputCustom<NewArticleFormValues>
+            noBrokenCorners
+            rules={{ required: true }}
+            height={48}
+            label="Preview title"
+            placeHolder="Type title here"
+            name="title"
+            control={newArticleForm.control}
+            variant="labelOutside"
+            containerStyle={{ marginVertical: layout.spacing_x3 }}
+            boxMainContainerStyle={{
+              backgroundColor: neutral00,
+              borderRadius: 12,
+            }}
           />
+
+          <TextInputCustom<NewArticleFormValues>
+            noBrokenCorners
+            rules={{ required: true }}
+            multiline
+            label="Preview subtitle"
+            placeHolder="Type short description here"
+            name="shortDescription"
+            control={newArticleForm.control}
+            variant="labelOutside"
+            containerStyle={{ marginBottom: layout.spacing_x3 }}
+            boxMainContainerStyle={{
+              backgroundColor: neutral00,
+              borderRadius: 12,
+            }}
+          />
+
+          <Label>Article preview</Label>
+          <SpacerColumn size={1.5} />
         </View>
+
+        <View
+          style={{
+            width: isSmallScreen ? windowWidth : width,
+            maxWidth: screenContentMaxWidth,
+            alignSelf: "center",
+          }}
+        >
+          <View>
+            <FileUploader
+              label="Thumbnail image"
+              onUpload={(files) =>
+                newArticleForm.setValue("thumbnailImage", {
+                  isThumbnailImage: true,
+                  ...files[0],
+                })
+              }
+              mimeTypes={IMAGE_MIME_TYPES}
+            >
+              {({ onPress }) => (
+                <CustomPressable
+                  onHoverIn={() => setThumbnailButtonHovered(true)}
+                  onHoverOut={() => setThumbnailButtonHovered(false)}
+                  onPress={onPress}
+                  style={{
+                    position: "absolute",
+                    right: 8,
+                    top: 8,
+                    zIndex: 1,
+                    backgroundColor: neutral00,
+                    borderColor: isThumbnailButtonHovered
+                      ? neutralFF
+                      : neutral33,
+                    borderWidth: 1,
+                    borderRadius: 999,
+                    height: 36,
+                    width: 36,
+                    justifyContent: "center",
+                    alignItems: "center",
+                  }}
+                >
+                  <SVG
+                    source={penSVG}
+                    width={24}
+                    height={24}
+                    color={neutralFF}
+                  />
+                </CustomPressable>
+              )}
+            </FileUploader>
+
+            <SocialArticleCard
+              style={cardStyle}
+              disabled
+              post={{
+                category: PostCategory.MarkdownArticle,
+                isDeleted: false,
+                identifier: "",
+                metadata: JSON.stringify(previewMetadata),
+                parentPostIdentifier: "",
+                subPostLength: 0,
+                authorId: userId || "",
+                createdAt: Date.parse("2024-06-17") / 1000,
+                reactions: [],
+                tipAmount: 0,
+                premiumLevel: 0,
+                id: "",
+                localIdentifier: "",
+                networkId: selectedNetworkId,
+              }}
+            />
+          </View>
+        </View>
+        <SpacerColumn size={3} />
+
+        <FormProvider {...newArticleForm}>
+          <ArticleContentEditor width={width} />
+        </FormProvider>
 
         {step.id !== "UNDEFINED" && isProgressBarShown && (
           <>
