+++ packages/screens/FeedNewArticle/components/ArticleContentEditor/ArticleContentEditor.tsx
@@ -1,5 +1,6 @@
 import type { MixedStyleRecord } from "@native-html/transient-render-engine";
 import markdownit from "markdown-it";
+import { full as emoji } from "markdown-it-emoji";
 import { FC, useRef, useState } from "react";
 import { Controller, useFormContext } from "react-hook-form";
 import {
@@ -18,10 +19,15 @@ import { useMaxResolution } from "@/hooks/useMaxResolution";
 import { Toolbar } from "@/screens/FeedNewArticle/components/ArticleContentEditor/Toolbar/Toolbar";
 import { ContentMode } from "@/screens/FeedNewArticle/components/ArticleContentEditor/utils";
 import { ARTICLE_MAX_WIDTH } from "@/utils/social-feed";
-import { neutral00, neutralA3, neutralFF } from "@/utils/style/colors";
+import {
+  neutral00,
+  neutral17,
+  neutralA3,
+  neutralFF,
+  primaryColor,
+} from "@/utils/style/colors";
 import { layout, RESPONSIVE_BREAKPOINT_S } from "@/utils/style/layout";
 import { NewArticleFormValues } from "@/utils/types/feed";
-import { LocalFileData } from "@/utils/types/files";
 
 interface Props {
   width: number;
@@ -34,12 +40,9 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
   const textInputRef = useRef<TextInput>(null);
   const [isTextInputHovered, setTextInputHovered] = useState(false);
   const borderWidth = 1;
-  const textInputMinHeight = height - 140;
+  const textInputMinHeight = height - 142;
+  const editionAndPreviewHeight = height - 80;
   const [textInputHeight, setTextInputHeight] = useState(textInputMinHeight);
-  const [textInputSelection, setTextInputSelection] = useState({
-    start: 0,
-    end: 0,
-  }); // Used to get the cursor position
   const [mode, setMode] = useState<ContentMode>("BOTH");
   const [renderWidth, setRenderWidth] = useState(0);
   const textInputContainerPadding =
@@ -47,47 +50,115 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
       ? 0
       : layout.spacing_x2 - borderWidth * 2;
 
-  // TODO: Find type: evt.nativeEvent.target was not recognised by LayoutChangeEvent, but the object target exist, so i don't understand (Same as NewsFeedInput)
-  // https://github.com/necolas/react-native-web/issues/795#issuecomment-1297511068, fix that i found for shrink lines when we deleting lines in the input
-  const adjustTextInputSize = (evt: any) => {
-    const el = evt?.nativeEvent?.target;
-    if (el) {
-      el.style.height = 0;
-      const newHeight = el.offsetHeight - el.clientHeight + el.scrollHeight;
-      el.style.height = `${newHeight < textInputMinHeight ? textInputMinHeight : newHeight}px`;
-    }
-  };
-
   // ========== Form
-  const { watch, control, setValue } = useFormContext<NewArticleFormValues>();
+  const { watch, control } = useFormContext<NewArticleFormValues>();
   const message = watch("message");
 
-  const addImage = (file: LocalFileData) => {
-    if (!file?.url) {
-      console.error("Invalid file: Missing URL or MIME type");
-      return;
-    }
-    const textBefore = message.substring(0, textInputSelection.start);
-    const textAfter = message.substring(
-      textInputSelection.start,
-      message.length - 1,
-    );
-    const result = `${textBefore}![image](${file.url})${textAfter}`;
-    setValue("message", result);
-  };
-
   // ========== Markdown
-  const md = markdownit();
+  const md = markdownit({
+    linkify: true,
+    breaks: true,
+  }).use(emoji);
   const html = md.render(message);
 
-  // TODO: Style this
   const markdownTagStyles: MixedStyleRecord = {
-    body: { fontSize: 14, color: neutralA3 },
-    h1: { fontSize: 24, fontWeight: "bold" },
-    p: { margin: 0 },
-    strong: { fontWeight: "bold", color: "yellow" },
-    a: { color: "blue", textDecorationLine: "underline" },
-    // img: {resizeMode: "contain"}
+    body: {
+      color: neutralA3,
+      fontSize: 14,
+      letterSpacing: -(14 * 0.04),
+      lineHeight: 22,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    p: {
+      marginVertical: 4,
+      color: neutralA3,
+      fontSize: 14,
+      letterSpacing: -(14 * 0.04),
+      lineHeight: 22,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    code: {
+      color: neutralA3,
+      fontSize: 13,
+      letterSpacing: -(13 * 0.04),
+      lineHeight: 22,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+
+      backgroundColor: neutral17,
+      marginVertical: 4,
+      paddingHorizontal: 4,
+      paddingVertical: 2,
+      borderRadius: 4,
+      alignSelf: "flex-start",
+    },
+    pre: {
+      fontSize: 13,
+      letterSpacing: -(13 * 0.04),
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+
+      backgroundColor: neutral17,
+      paddingHorizontal: 8,
+      paddingVertical: 8,
+      borderRadius: 4,
+    },
+    strong: { fontWeight: "700" },
+    a: {
+      color: primaryColor,
+      textDecorationLine: "none",
+    },
+    hr: { backgroundColor: neutralA3 },
+    h1: {
+      color: neutralFF,
+      fontSize: 28,
+      letterSpacing: -(28 * 0.02),
+      lineHeight: 37,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    h2: {
+      color: neutralFF,
+      fontSize: 21,
+      letterSpacing: -(21 * 0.02),
+      lineHeight: 28,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    h3: {
+      color: neutralFF,
+      fontSize: 16,
+      letterSpacing: -(16 * 0.02),
+      lineHeight: 23,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    h4: {
+      color: neutralFF,
+      fontSize: 14,
+      letterSpacing: -(14 * 0.04),
+      lineHeight: 20,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    h5: {
+      color: neutralA3,
+      fontSize: 14,
+      letterSpacing: -(14 * 0.04),
+      lineHeight: 20,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
+    h6: {
+      color: neutralA3,
+      fontSize: 12,
+      letterSpacing: -(12 * 0.04),
+      lineHeight: 16,
+      fontFamily: "Exo_500Medium",
+      fontWeight: "500",
+    },
   };
 
   // ========== JSX
@@ -99,22 +170,7 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
       }}
     >
       {/* ==== Toolbar */}
-      <Toolbar
-        addEmoji={() => {
-          /*TODO*/
-        }}
-        addGIF={() => {
-          /*TODO*/
-        }}
-        addAudio={() => {
-          /*TODO*/
-        }}
-        addVideo={() => {
-          /*TODO*/
-        }}
-        addImage={addImage}
-        setMode={setMode}
-      />
+      <Toolbar setMode={setMode} />
       <SpacerColumn size={2} />
 
       {/* ==== Edition and preview */}
@@ -122,8 +178,8 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
         style={{
           flexDirection: "row",
           width: windowWidth < RESPONSIVE_BREAKPOINT_S ? "100%" : width,
-          maxWidth: ARTICLE_MAX_WIDTH + layout.spacing_x2 * 2,
-          // height
+          maxWidth: ARTICLE_MAX_WIDTH + 16 * 2,
+          height: editionAndPreviewHeight,
         }}
       >
         {/* ==== Edition */}
@@ -153,9 +209,11 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
                   onChange: (value: string) => void;
                 };
                 return (
-                  <View
+                  <ScrollView
+                    showsHorizontalScrollIndicator={false}
                     style={[
                       {
+                        flex: 1,
                         padding: textInputContainerPadding,
                         borderRadius: 12,
                       },
@@ -166,6 +224,7 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
                     ]}
                   >
                     <TextInput
+                      scrollEnabled={false}
                       multiline
                       value={value}
                       style={[
@@ -174,21 +233,17 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
                           outlineStyle: "none",
                           color: neutralA3,
                           border: "none",
+                          textAlignVertical: "top",
                         } as TextStyle,
                       ]}
-                      onChange={adjustTextInputSize}
-                      onLayout={adjustTextInputSize}
                       onChangeText={onChange}
-                      onSelectionChange={(e) =>
-                        setTextInputSelection(e.nativeEvent.selection)
-                      }
                       onContentSizeChange={(e) => {
                         // The input grows depending on the content height
                         setTextInputHeight(e.nativeEvent.contentSize.height);
                       }}
                       ref={textInputRef}
                     />
-                  </View>
+                  </ScrollView>
                 );
               }}
             />
@@ -215,12 +270,6 @@ export const ArticleContentEditor: FC<Props> = ({ width }) => {
                 paddingHorizontal: textInputContainerPadding,
                 marginTop: textInputContainerPadding,
               }}
-              contentContainerStyle={[
-                mode !== "PREVIEW" && {
-                  height: textInputHeight,
-                },
-                { minHeight: textInputMinHeight },
-              ]}
               onLayout={(e) => setRenderWidth(e.nativeEvent.layout.width)}
             >
               <RenderHtml
