+++ packages/utils/navigation.ts
@@ -61,12 +61,6 @@ export type RootStackParamList = {
         additionalHashtag?: string;
       })
     | undefined;
-  FeedNewArticleMarkdown:
-    | (NewPostFormValues & {
-        additionalMention?: string;
-        additionalHashtag?: string;
-      })
-    | undefined;
   FeedPostView: { id: string };
   HashtagFeed: { hashtag: string };
 
@@ -245,7 +239,6 @@ const getNavConfig: (homeScreen: keyof RootStackParamList) => NavConfig = (
       NFTDetail: "nft/:id",
       Feed: "feed/:tab?",
       FeedNewArticle: "feed/new",
-      FeedNewArticleMarkdown: "feed/new-md",
       FeedPostView: "feed/post/:id",
       HashtagFeed: "feed/tag/:hashtag",
